Now conceptually correct working console app
App plays mastermind to itself
It first makes up a random colorlist: the truth
It than goes through all permuations for a colorlist of 4 colors that
can be chosen from 6 colors.
For each permuation it verifies it against the knowlagde base. If it is
in line with the knowledge base, the permuation is accepted as a
hypothesis and put on the board. The permutation is than judged and
added to the knowledge base. Is the judgement is a list of 4 times
Black, the game ends.